3Match 시 블록 삭제 & 블록 떨어지는 애니메이션 & 블록 생성 애니메이션

+++ Match3Puzzle/Classes/SceneIngame.cpp
@@ -37,9 +37,23 @@ void SceneIngame::setBlockSprite(int x, int y, Sprite * s)
 void SceneIngame::destroyBlock(int x, int y)
 {
 	if (blockData[y][x] != 0) {
-		blockSprite[y][x]->removeFromParent();
+		state = GameState::BLOCK_MOVING;
+		blockSprite[y][x]->runAction(Sequence::create(
+			FadeOut::create(0.125f),
+			FadeIn::create(0.125f),
+			FadeOut::create(0.125f),
+			FadeIn::create(0.125f),
+			Spawn::create(ScaleTo::create(0.125f, 0.0), FadeOut::create(0.125f), nullptr),
+			RemoveSelf::create(),
+			nullptr));
 		blockSprite[y][x] = nullptr;
 		blockData[y][x] = 0;
+
+		this->runAction(Sequence::create(
+			DelayTime::create(0.625f),
+			CallFunc::create([=]() {dropBlocks(x); }),
+			nullptr
+			));
 	}
 }
 
@@ -84,11 +98,17 @@ int SceneIngame::findFilledBlockYIndex(int x, int y)
 
 void SceneIngame::dropBlocks(int x)
 {
+	bool isDrop = false;
 	for (int i = 0; i < BLOCK_VERTICAL; i++) {
 		int empty_y = findEmptyBlockYIndex(x, i);
-
+		if (empty_y == -1) continue;
 		int filled_y = findFilledBlockYIndex(x, empty_y);
-
+		if (filled_y == -1) { 
+			createBlock(x, empty_y, rand() % BLOCK_VAR + 1);
+			blockSprite[empty_y][x]->setPosition(convertBlockCoordToGameCoord(Vec2(x, BLOCK_VERTICAL + 1)));
+			blockSprite[empty_y][x]->runAction(MoveTo::create(0.125f, convertBlockCoordToGameCoord(Vec2(x, empty_y))));
+			continue; 
+		}
 		{
 			int a = getBlockData(x, empty_y);
 			int b = getBlockData(x, filled_y);
@@ -101,10 +121,103 @@ void SceneIngame::dropBlocks(int x)
 			SWAP(Sprite*, a, b);
 			setBlockSprite(x, empty_y, a);
 			setBlockSprite(x, filled_y, b);
+
+			a->stopAllActions();
+			a->runAction(MoveTo::create(0.125f, convertBlockCoordToGameCoord(Vec2(x, empty_y))));
+
 		}
-		alignBlockSprite();
+		isDrop = true;
+	}
+	if (isDrop) {
+		for (int i = 0; i < BLOCK_VERTICAL; i++) {
+			judgeMatch(x, i);
+		}
+	}
+	else {
+		state = GameState::PLAYING;
 	}
 
+}
+
+void SceneIngame::stackPush(Vec2 value)
+{
+	if (judgeData[(int)value.y][(int)value.x] != 0) return;
+	judgeStack[judgeStackCount++] = value;
+	judgeData[(int)value.y][(int)value.x] = 1;
+
+}
+
+Vec2 SceneIngame::stackPop()
+{
+	auto ret = judgeStack[--judgeStackCount];
+	judgeData[(int)ret.y][(int)ret.x] = 0;
+	return ret;
+}
+
+void SceneIngame::stackEmpty()
+{
+	judgeStackCount = 0;
+	for (int i = 0; i < BLOCK_HORIZONTAL; i++) {
+		for (int k = 0; k < BLOCK_VERTICAL; k++) {
+			judgeData[k][i] = 0;
+		}
+	}
+}
+
+bool SceneIngame::stackFind(Vec2 value)
+{
+	return judgeData[(int)value.y][(int)value.x] == 1;
+}
+
+void SceneIngame::judgeMatch(int x, int y)
+{
+	int blockData = getBlockData(x, y);
+	if (blockData == 0)return;
+
+	stackPush(Vec2(x, y));
+	int push_cnt = 0;
+	for (int i = 0; i < 4; i++) {
+		int next_x = x;
+		int next_y = y;
+		int inc_x;
+		int inc_y;
+
+		switch (i) {
+		case 0: inc_x = 1; inc_y = 0; push_cnt = 0; break;
+		case 1: inc_x = -1; inc_y = 0; break;
+		case 2: inc_x = 0; inc_y = 1; push_cnt = 0; break;
+		case 3: inc_x = 0; inc_y = -1; break;
+		}
+
+		while (true) {
+			next_x += inc_x;
+			next_y += inc_y;
+			if (next_x < 0 || next_x >= BLOCK_HORIZONTAL) break;
+			if (next_y < 0 || next_y >= BLOCK_VERTICAL) break;
+			if (getBlockData(next_x, next_y) == blockData) {
+				stackPush(Vec2(next_x, next_y));
+				push_cnt++;
+			}
+			else break;
+		}
+
+		if (i % 2 == 0) continue;
+		if (push_cnt < 2) {
+			for (int k = 0; k < push_cnt; k++) {
+				stackPop();
+			}
+		}
+	}
+	if (judgeStackCount > 1) {
+		while (judgeStackCount > 0) {
+			Vec2 p = stackPop();
+			destroyBlock(p.x, p.y);
+		}
+	}
+	else {
+		state = GameState::PLAYING;
+	}
+	stackEmpty();
 }
 
 SceneIngame * SceneIngame::create()
@@ -152,7 +265,7 @@ void SceneIngame::initGame()
 {
 	for (int i = 0; i < BLOCK_HORIZONTAL; i++) {
 		for (int k = 0; k < BLOCK_VERTICAL; k++) {
-			createBlock(i, k, rand()%4 + 1);
+			createBlock(i, k, rand()%BLOCK_VAR + 1);
 		}
 	}
 
@@ -180,9 +293,19 @@ void SceneIngame::alignBlockSprite()
 bool SceneIngame::onTouchBegan(Touch * t, Event * e)
 {
 	Vec2 p = convertGameCoordToBlockCoord(t->getLocation());
-	CCLOG("%f, %f", p.x, p.y);
-	destroyBlock(p.x, p.y);
-	dropBlocks(p.x);
+
+	if (state == GameState::PLAYING) {
+		if (p.x >= BLOCK_HORIZONTAL || p.x < 0) return true;
+		if (p.y >= BLOCK_VERTICAL || p.y < 0) return true;
+
+		CCLOG("%f, %f", p.x, p.y);
+		destroyBlock(p.x, p.y);
+
+		for (int i = 0; i < BLOCK_VERTICAL; i++) {
+			judgeMatch(p.x, i);
+		}
+	}
+
 
 	return true;
 }
